Replace nanoid() calls with crypto.randomUUID()
Motivation:
The prod env on Fly.io uses a Postgres DB, which appears to have a UUID type. Thus, changes sent to the database fail because nanoids don't match UUIDs.

Also, on the bright side, it seems UUIDs are suited for distributed systems (read the Postgres docs on UUID for example), so I can use them without worry.